Refactor enemy components, decouple from EventBus and optimize UI (#416)
