+++ Assets/Prefab/Enemy/ZombiBody.prefab
@@ -76,16 +76,10 @@ PrefabInstance:
       addedObject: {fileID: 3090767071386169047}
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 725b184d2fae6684f84ada0fae3af288, type: 3}
       insertIndex: -1
-      addedObject: {fileID: 6298201036355012212}
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 725b184d2fae6684f84ada0fae3af288, type: 3}
-      insertIndex: -1
       addedObject: {fileID: 5421981319966377777}
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 725b184d2fae6684f84ada0fae3af288, type: 3}
       insertIndex: -1
       addedObject: {fileID: 3699580084391280425}
-    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 725b184d2fae6684f84ada0fae3af288, type: 3}
-      insertIndex: -1
-      addedObject: {fileID: 1508460088435983056}
   m_SourcePrefab: {fileID: 100100000, guid: 725b184d2fae6684f84ada0fae3af288, type: 3}
 --- !u!1 &8324395216365602149 stripped
 GameObject:
@@ -106,7 +100,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: Assembly-CSharp::Game.Core.Enemy.EnemyHitReceiver
   _minimumHitSpeed: 0.75
   _bodyDamageMultiplier: 2.5
-  _sameItemCooldown: 0.25
   _despawnItemOnHit: 1
 --- !u!95 &3090767071386169047
 Animator:
@@ -130,21 +123,6 @@ Animator:
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorStateOnDisable: 0
   m_WriteDefaultValuesOnDisable: 0
---- !u!114 &6298201036355012212
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8324395216365602149}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c3a4985fbfc99334686a09c5a519b32b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: Assembly-CSharp::EnemyAnimation
-  _animator: {fileID: 3090767071386169047}
-  _maxHitAnimationTime: 1
-  _addHitAnimationTime: 1
 --- !u!114 &5421981319966377777
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -157,6 +135,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 71fa6540b0f3d294abbeac213ab6ae93, type: 3}
   m_Name: 
   m_EditorClassIdentifier: Assembly-CSharp::EnemyBodyController
+  _animator: {fileID: 3090767071386169047}
+  _maxHitAnimationTime: 1
+  _addHitAnimationTime: 1
+  _dropDuration: 0.3
+  _dropRot: {x: -180, y: 0, z: 0}
 --- !u!136 &3699580084391280425
 CapsuleCollider:
   m_ObjectHideFlags: 0
@@ -180,17 +163,3 @@ CapsuleCollider:
   m_Height: 1
   m_Direction: 1
   m_Center: {x: 0, y: 0.2, z: 0}
---- !u!114 &1508460088435983056
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8324395216365602149}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8b0ff27d9e1e35140957f6f728868003, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: Assembly-CSharp::EnemyBodyPose
-  _dropDuration: 0.3
-  _dropRot: {x: -180, y: 0, z: 0}
