--- Assets/Scripts/Core/Enemy/BarrierController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a969eb88586cb684c940001ea1d240ec
