--- Assets/Scripts/Core/Enemy/BaseHitRecovery.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7053aa2824c77474284a9713a483ff37
