--- Assets/Scripts/Core/Enemy/EnemyAnimation.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: c3a4985fbfc99334686a09c5a519b32b
