--- Assets/Scripts/Core/Enemy/EnemyAttack.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 68041e764397d1c4693f55bd6bcfea95
