--- Assets/Scripts/Core/Enemy/EnemyAttackGauge.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7c3e431e464f0e148ab2cc43c18642f1
