--- Assets/Scripts/Core/Enemy/EnemyBodyPose.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8b0ff27d9e1e35140957f6f728868003
