--- Assets/Scripts/Core/Enemy/EnemyHoldCounter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b9120970249b3bf4fb81b738bcdbe9b8
