--- Assets/Scripts/Core/Enemy/EnemyStateManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 332ea97ffddfb3b4db56c95d8a8f56c5
