--- Assets/Scripts/Core/Enemy/IBarrier.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d63017848eb908a479da3a5038ceccd8
